bug fix #779

+++ TreeConverter/TreeConversion/syntax_tree_visitor.cs
@@ -1206,14 +1206,14 @@ namespace PascalABCCompiler.TreeConverter
                     bfc = ((right as typed_expression).type as delegated_methods).proper_methods[0];
                     right = convertion_data_and_alghoritms.explicit_convert_type(right, CreateDelegate(bfc.simple_function_node));
                     sil2 = right.type.find_in_type(name);
-                    if (saved_sil != null)
+                    if (saved_sil != null && sil != null)
                     {
                         sil.RemoveRange(1, sil.Count() - 1);
                         sil.Add(saved_sil);
                     }
                     else
                         saved_sil = sil;
-                    if (saved_sil2 != null)
+                    if (saved_sil2 != null && sil2 != null)
                     {
                         sil2.RemoveRange(1, sil2.Count() - 1);
                         sil2.Add(saved_sil2);
